requirements.txt
django
flake8
pytest
black
pillow
djangorestframework
django-cors-headers
